Some non-functional garbage from which someone may derive something useful
It worked once but the plugin code could not be updated/changed and poetry refuses to reinstall it so I'm stepping back for a bit. It did successfully demo SDAP adding in new plugin algs for packages added in web.ini, but I can't seem to get it to be reliably importable.